fix: use correct perimeter style property in stories (#316)
Some stories used a `perimiter` property instead of `perimeter` (i
instead of e). This is now fixed. When possible, do not set the
RectanglePerimeter in the style configuration as it is already set in
the default vertex style.
Migrate some stories to TypeScript when possible (migrating Swimlanes
involved too many unrelated changes, so it will be done later).
The migration to TS required to remove the usage of constants to avoid
errors like "Cannot access ambient const enums when 'isolatedModules' is
enabled".

The following stories already had issues unrelated to the perimeter
property and are still not working. This is why they weren't migrated to
TypeScript as part of the fix.
  - FileIO
  - Monitor
  - Scrollbars