fix(stories): display images when not serving in the root context (#340)
This prepares the deployment to GitHub Pages. Previously, images were
correctly served only when Storybook was deployed in the root context,
for example when accessing http://localhost:8080 but not for
http://localhost:8080/maxGraph.
The problem was due to the usage of leading / in images path. Some paths
that involved "editors" images (not available) have been fixed as well.

The creation of the "div" container used to display the Graph is now
managed at a single place to remove duplications.

Additional fix in SwimLanes.stories.js to restore some restore mxGraph
behaviors:
  - no fill color
- no background grid image The story is still not working as in mxGraph,
mainly because it doesn't use the Editor and key bindings defined in a
xml configuration file that was provided in mxGraph examples.