refactor: migrate the DynamicToolbar story to TypeScript (#341)
Fix the style to ensure that a dropped element uses the correct shape.
Previously, all dropped elements used the Rectangle shape instead of the
shape chosen in the toolbar. Also add a perimeter to the shapes to
better handle the rendering of connections.
Remove an additional listener previously used for IE in mxGraph, as IE
is not supported in maxGraph.
Comment the use of the KeyHandler, which is currently not working.

Improve the page layout: better align the toolbar elements.