chore!: build with TypeScript v5.3.3 (#279)
As TypeScript v5 detects more errors, the version bump requires to
update the `Multiplicity` class. It now only accepts number for the
`max` property. Previously, it accepted a special string value, but
TypeScript v5 now complains when comparing a number with a string. So,
as the `max` property is used in number comparisons, it must only accept
number.

Also bump some development dependencies to get better support for
TypeScript v5.
  - `typedoc` from 0.23 to 0.25 
  - `storybook` from 7.4  to 7.6

BREAKING CHANGE: it is no longer possible to pass an 'n' value for the
`max` property of the `Multiplicity` class. Pass `null` instead to have
the same effect.