fix: make usage of Codecs work in stories (#350)
Previously, some stories still did encode and/or decode by calling
directly the Codec code. Default Codecs were not registered so errors
occurred. This is now done by using `ModelXmlSerializer`.
Most of the stories updated as part of the fix have been migrated to
TypesScript to also fix other minor issues.

Clipboard (still in JavaScript)
The Clipboard actions still don't work and have not been touched as this
is out of the scope of the changes proposed here.
The code that uses the Codecs to generate the XML string used for
copy/paste (the string representation of the cells stored in the
clipboard) is now working (as the Codecs are registered).

FileIO
  - load of txt and xml file works
- the user can choose the source of the model (text or xml files) with a
Storybook control
  - display images
  - display the "arrange" button

JSonData
  - display the content
  - display the popup

UserObject
  - load images (for popup)
  - fix style and form div
  - improve the layout
  - fix error when validating new value in the form (wrong import)

DISCLAIMER about the maxGraph "popup"
The CSS class are still not correctly set in the `maxGraph` code, so
popups still don't display correctly.
The popups are not cleaned when changing from a story to another, as
they are currently not created in the DOM at the right place. They are
created directly in whole document, so not at the place where the story
is located in the DOM (not in the right div).