fix: correctly update style when calling Graph methods (#380)
`GraphDataModel.setStyle` does the actual style change. Its current
implementation requires to pass an instance of CellStyle which is
different from the current style hold by the Cell. Previously, the Graph
methods directly update the Cell style instance prior calling the style
update so nothing happened.
The problem was spotted in the Monitor story which is now fixed.
Also fix the signature of Graph.setCellStyle as the `cells` parameter is
optional (previously, it was marked as mandatory).

Introduce a convenient `Cell.getClonedStyle` method to ease style update
and update the JSDOC of several functions to warn users about how to
correctly make the style update happen.

The fix has a slight side effects exhibited by the EdgeMixin tests. The
`sourcePort` and `targetPort` cell style properties are now set to null
at vertex insertion. This should not have any effects for the end user.