fix: correctly manage default `style.foldable` value (#385)
Previously, when a cell's `style.foldable` property was not set,
everything happened as if it was set to `false`, whereas it should have
considered as `true`. Thus, by default, the expand image was not
displayed, and it was not possible to expand the cell.

The problem was due to an error during mxGraph migration.


Some stories exhibited the problem. The expand/collapse images and
actions are now available.
  - `Folding`:
    - It has been migrated to TypeScript and updated to work better
- Some mxGraph features still don't work (highlighting the parent when
dragging a cell, parent association).
- `Swimlanes` and `Collapse`. Some mxGraph features still don't work
like applying custom style when the cell are collapsed