fix: add null check to connection validation message (#364)
Add a null check to the connection validation message in `ConnectionHandlerCellMarker` to handle invalid connection targets.
This is similar to the syntax used in mxGraph. 
It is specifically necessary to check for the presence of an empty string, as `validateConnection` returns an empty string to indicate that the connection target is invalid. 

A nullity check must also be added to `Multiplicity.checkTerminal`. This null error was visible in the Swimlanes story.

---------

Co-authored-by: Eric Redekopp <eric.redekopp@usask.ca>
Co-authored-by: Thomas Bouffard <27200110+tbouffard@users.noreply.github.com>