ci: include js-example in the "build" workflow (#402)
This ensures that errors are detected sooner.

Also bump webpack dependencies to use the latest webpack version.
Fix the configuration to make the js-example build pass
  - fix package-lock.json: the globby dependency wasn't found
  - update `webpack.config.js` to make webpack happy (this is a workaround). maxGraph is now
  declared as "type=module" but the import in the js files are not including extension.
```
ERROR in ../core/dist/index.js 192:0-24
Module not found: Error: Can't resolve './types' in '/dev/maxGraph/packages/core/dist'
Did you mean 'types.js'?
BREAKING CHANGE: The request './types' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.
```