fix: correctly set the default CSS class name for MaxWindow (#407)
Restore the class name used in mxGraph to increase compatibility.
Previously, the default value was empty.

Also update the CSS rules common.css to correctly use the `mxWindow`
classes.
This restores the original mxGraph CSS content which should ease the
migration.

Notice that this change allows to fix the stories involving `MaxWindow`:
  - HelloPort
  - JSonData
  - UserObjects
  - Windows

The Windows story has been migrated to TypeScript to ease the
maintenance.
The popup are now made closable. Currently, when switching from the
Windows story to another one, the popup are not destroyed and are still
displayed in the other story. Making them closable provide a visual
workaround to remove them manually.
Solving this problem is outside the scope of the changes proposed here.
It also has an impact on all stories using `MaxWindow` or other
additions of graphical elements to the graph (like `MaxPopupMenu`, for
example).