docs(release): attach examples and website (#435)
The live website is for the next version. As the API is not stable for
now and breaking changes occur, providing an archive of the website
(which includes the documentation for usage and API) let user access to
the documentation of the specific version of maxGraph that they are
using.
Providing a static archive of the example will help detecting eventual
regression, or see the evolution of the library.