feat!: introduce configurable logger (#431)
Previously, logs were sent to the browser console or to `MaxLog`, which
was inconsistent across the code base.
This is now configurable and maxGraph no longer produces logs by
default.

The logger is configured globally (i.e. it is not specific to a `Graph`
instance) and all the examples in the JSDoc have been updated to
demonstrate how to use the new logging system (instead of using `MaxLog`
directly).
It is also possible to provide your own implementation of the logger.
Please note that this is a first implementation and may change in the
future.

This change also improves tree-shaking. `MaxLog` is no longer included
in applications using maxGraph, and transitively, `MaxWindow` is not
included either.

Sending logs to `MaxLog` by default also posed a problem in terms of
user interface.
For example, if an error occurred in `ConnectionHandler`, the `MaxLog`
window opened in the application including maxGraph, whereas this was
probably not what users wanted.
In addition, if maxGraph's CSS rules were not available (for example, by
importing the CSS file supplied by maxGraph), the `MaxLog` window was
not displayed correctly: its content was added to the bottom of the page
without any styling.

Notes on `ConsoleLogger` logs
By default, only warnings and errors are sent to the browser console to
reduce the amount of logs. It is possible to activate info, debug and
trace. Remember that traces are produced each time the model is
validated.

No direct console logs allowed in maxGraph code
A new eslint rule has been configured to prohibit the use of console
logs, as the global log system must be used instead.

Impact on the `Window` story
This is the only story that explicitly uses `MaxLog`, so it is
configured to direct maxGraph logs to `MaxLog`. To avoid side-effects on
other stories because the logger configuration is global, the storybook
configuration has been set to reset the logger to its default value
before initializing new stories.

BREAKING CHANGES:
- Logs are no longer sent to `MaxLog` by default. To restore the
previous behavior, change maxGraph's global configuration with:
`GlobalConfig.logger = new MaxLogAsLogger();`
- `MaxWindow.activeWindow` is no longer available; it was intended for
internal use only, so there's no reason to make it public.