chore(test): switch from ts-jest to @swc/jest (#540)
This makes the whole test run faster because `swc` uses native binaries
to transpile the TypeScript code.
This is particularly noticeable when you run a single test in your IDE.

The gain (i.e. the difference between execution with swc and with
ts-jest) is very important: from 21 seconds to 6 seconds!

Unlike `ts-jest`, `swc` does no type checking (it just transpiles). A
new npm script is available to type-check the test code.