fix(label): improve editing position and 'label changed" event (#567)
The label editing area for both vertices and edges is now properly aligned with the label's position in "view" mode. Previously, it was always incorrectly placed at the bottom of the page.

Additionally, the event passed to the `Graph.labelChanged` method is now correctly handled. Previously, the event was `null`. While this did not affect maxGraph's default implementation, it could impact customizations relying on the event to manage updates to the `Cell` value.  
This fix is demonstrated in the `Editing` story, which utilizes this feature.